IS-1663: Refactor tests
