+++ src/test/kotlin/no/nav/syfo/aktivitetskrav/api/AktivitetskravApiForhandsvarselSpek.kt
@@ -1,0 +1,187 @@
+package no.nav.syfo.aktivitetskrav.api
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.module.kotlin.readValue
+import io.ktor.http.*
+import io.ktor.server.testing.*
+import io.mockk.*
+import no.nav.syfo.aktivitetskrav.AktivitetskravService
+import no.nav.syfo.aktivitetskrav.database.AktivitetskravRepository
+import no.nav.syfo.aktivitetskrav.domain.*
+import no.nav.syfo.aktivitetskrav.kafka.AktivitetskravVurderingProducer
+import no.nav.syfo.aktivitetskrav.kafka.domain.KafkaAktivitetskravVurdering
+import no.nav.syfo.domain.PersonIdent
+import no.nav.syfo.testhelper.*
+import no.nav.syfo.testhelper.generator.createAktivitetskravForTest
+import no.nav.syfo.testhelper.generator.createAktivitetskravNy
+import no.nav.syfo.testhelper.generator.generateDocumentComponentDTO
+import no.nav.syfo.util.*
+import org.amshove.kluent.*
+import org.apache.kafka.clients.producer.KafkaProducer
+import org.apache.kafka.clients.producer.RecordMetadata
+import org.spekframework.spek2.Spek
+import org.spekframework.spek2.style.specification.describe
+import java.time.LocalDate
+import java.util.*
+import java.util.concurrent.Future
+
+class AktivitetskravApiForhandsvarselSpek : Spek({
+    val objectMapper: ObjectMapper = configuredJacksonMapper()
+    val urlAktivitetskravPerson = "$aktivitetskravApiBasePath/$aktivitetskravApiPersonidentPath"
+
+    val nyAktivitetskrav = createAktivitetskravNy(
+        tilfelleStart = LocalDate.now().minusWeeks(10),
+    ).copy(
+        createdAt = nowUTC().minusWeeks(2)
+    )
+
+    describe(AktivitetskravApiForhandsvarselSpek::class.java.simpleName) {
+        with(TestApplicationEngine()) {
+            start()
+            val externalMockEnvironment = ExternalMockEnvironment.instance
+            val database = externalMockEnvironment.database
+            val kafkaProducer = mockk<KafkaProducer<String, KafkaAktivitetskravVurdering>>()
+
+            application.testApiModule(
+                externalMockEnvironment = externalMockEnvironment,
+                aktivitetskravVurderingProducer = AktivitetskravVurderingProducer(
+                    producer = kafkaProducer,
+                ),
+            )
+            val aktivitetskravRepository = AktivitetskravRepository(database)
+            val aktivitetskravService = AktivitetskravService(
+                aktivitetskravRepository = aktivitetskravRepository,
+                aktivitetskravVurderingProducer = mockk(relaxed = true),
+                database = database,
+                arenaCutoff = externalMockEnvironment.environment.arenaCutoff,
+            )
+            val validToken = generateJWT(
+                audience = externalMockEnvironment.environment.azure.appClientId,
+                issuer = externalMockEnvironment.wellKnownInternalAzureAD.issuer,
+                navIdent = UserConstants.VEILEDER_IDENT,
+            )
+
+            beforeEachTest {
+                clearMocks(kafkaProducer)
+                coEvery {
+                    kafkaProducer.send(any())
+                } returns mockk<Future<RecordMetadata>>(relaxed = true)
+            }
+            afterEachTest {
+                database.dropData()
+            }
+
+            fun postForhandsvarsel(
+                aktivitetskravUuid: UUID = nyAktivitetskrav.uuid,
+                arbeidstakerPersonIdent: PersonIdent = UserConstants.ARBEIDSTAKER_PERSONIDENT,
+                newForhandsvarselDTO: ForhandsvarselDTO,
+            ) = run {
+                val url = "$aktivitetskravApiBasePath/${aktivitetskravUuid}$forhandsvarselPath"
+                handleRequest(HttpMethod.Post, url) {
+                    addHeader(HttpHeaders.ContentType, ContentType.Application.Json.toString())
+                    addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
+                    addHeader(NAV_PERSONIDENT_HEADER, arbeidstakerPersonIdent.value)
+                    setBody(objectMapper.writeValueAsString(newForhandsvarselDTO))
+                }
+            }
+
+            fun postAvvent(
+                aktivitetskravUuid: UUID = nyAktivitetskrav.uuid,
+                arbeidstakerPersonIdent: PersonIdent = UserConstants.ARBEIDSTAKER_PERSONIDENT,
+            ) = run {
+                val url = "$aktivitetskravApiBasePath/${aktivitetskravUuid}$vurderAktivitetskravPath"
+                handleRequest(HttpMethod.Post, url) {
+                    addHeader(HttpHeaders.ContentType, ContentType.Application.Json.toString())
+                    addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
+                    addHeader(NAV_PERSONIDENT_HEADER, arbeidstakerPersonIdent.value)
+                    setBody(
+                        objectMapper.writeValueAsString(
+                            AktivitetskravVurderingRequestDTO(
+                                status = AktivitetskravStatus.AVVENT,
+                                beskrivelse = "venter litt",
+                                arsaker = listOf(VurderingArsak.ANNET),
+                            )
+                        )
+                    )
+                }
+            }
+            describe("Forhåndsvarsel") {
+                beforeEachTest { aktivitetskravService.createAktivitetskravForTest(nyAktivitetskrav) }
+                val fritekst = "Dette er et forhåndsvarsel"
+                val forhandsvarselDTO = ForhandsvarselDTO(
+                    fritekst = fritekst,
+                    document = generateDocumentComponentDTO(
+                        fritekst = fritekst,
+                        header = "Forhåndsvarsel"
+                    )
+                )
+                describe("Happy path") {
+                    it("Successfully creates a new forhandsvarsel") {
+                        with(postForhandsvarsel(newForhandsvarselDTO = forhandsvarselDTO)) {
+                            response.status() shouldBeEqualTo HttpStatusCode.Created
+                            val createdForhandsvarsel =
+                                objectMapper.readValue(response.content, AktivitetskravVarsel::class.java)
+                            createdForhandsvarsel.document shouldBeEqualTo forhandsvarselDTO.document
+                        }
+
+                        with(
+                            handleRequest(HttpMethod.Get, urlAktivitetskravPerson) {
+                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
+                                addHeader(NAV_PERSONIDENT_HEADER, UserConstants.ARBEIDSTAKER_PERSONIDENT.value)
+                            }
+                        ) {
+                            response.status() shouldBeEqualTo HttpStatusCode.OK
+
+                            val responseDTOList =
+                                objectMapper.readValue<List<AktivitetskravResponseDTO>>(response.content!!)
+                            val aktivitetskravResponseDTO = responseDTOList.first()
+                            val varselResponseDTO = aktivitetskravResponseDTO.vurderinger.first().varsel
+                            varselResponseDTO.shouldNotBeNull()
+                            varselResponseDTO.svarfrist shouldBeEqualTo varselResponseDTO.createdAt.toLocalDate()
+                                .plusWeeks(3)
+                        }
+                    }
+                }
+
+                describe("Unhappy path") {
+                    it("Can't find aktivitetskrav for given uuid") {
+                        with(
+                            postForhandsvarsel(
+                                aktivitetskravUuid = UUID.randomUUID(),
+                                newForhandsvarselDTO = forhandsvarselDTO,
+                            )
+                        ) {
+                            response.status() shouldBeEqualTo HttpStatusCode.BadRequest
+                        }
+                    }
+
+                    it("Fails if document is empty") {
+                        val varselWithoutDocument = forhandsvarselDTO.copy(document = emptyList())
+                        with(postForhandsvarsel(newForhandsvarselDTO = varselWithoutDocument)) {
+                            response.status() shouldBeEqualTo HttpStatusCode.BadRequest
+                        }
+                    }
+                    it("Fails if already forhandsvarsel") {
+                        with(postForhandsvarsel(newForhandsvarselDTO = forhandsvarselDTO)) {
+                            response.status() shouldBeEqualTo HttpStatusCode.Created
+                        }
+                        with(postForhandsvarsel(newForhandsvarselDTO = forhandsvarselDTO)) {
+                            response.status() shouldBeEqualTo HttpStatusCode.BadRequest
+                        }
+                    }
+                    it("Fails if already forhandsvarsel before") {
+                        with(postForhandsvarsel(newForhandsvarselDTO = forhandsvarselDTO)) {
+                            response.status() shouldBeEqualTo HttpStatusCode.Created
+                        }
+                        with(postAvvent()) {
+                            response.status() shouldBeEqualTo HttpStatusCode.OK
+                        }
+                        with(postForhandsvarsel(newForhandsvarselDTO = forhandsvarselDTO)) {
+                            response.status() shouldBeEqualTo HttpStatusCode.Created // TODO: Should give Bad Request
+                        }
+                    }
+                }
+            }
+        }
+    }
+})
