+++ src/test/kotlin/no/nav/syfo/testhelper/TestApplicationEngineUtils.kt
@@ -1,0 +1,65 @@
+package no.nav.syfo.testhelper
+
+import io.ktor.http.*
+import io.ktor.server.testing.*
+import no.nav.syfo.util.NAV_PERSONIDENT_HEADER
+import no.nav.syfo.util.bearerHeader
+import org.amshove.kluent.shouldBeEqualTo
+
+fun TestApplicationEngine.testMissingToken(url: String, httpMethod: HttpMethod) {
+    with(
+        handleRequest(httpMethod, url) {}
+    ) {
+        response.status() shouldBeEqualTo HttpStatusCode.Unauthorized
+    }
+}
+
+fun TestApplicationEngine.testMissingPersonIdent(
+    url: String,
+    validToken: String,
+    httpMethod: HttpMethod,
+) {
+    with(
+        handleRequest(httpMethod, url) {
+            addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
+        }
+    ) {
+        response.status() shouldBeEqualTo HttpStatusCode.BadRequest
+    }
+}
+
+fun TestApplicationEngine.testInvalidPersonIdent(
+    url: String,
+    validToken: String,
+    httpMethod: HttpMethod,
+) {
+    with(
+        handleRequest(httpMethod, url) {
+            addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
+            addHeader(
+                NAV_PERSONIDENT_HEADER,
+                UserConstants.ARBEIDSTAKER_PERSONIDENT.value.drop(1)
+            )
+        }
+    ) {
+        response.status() shouldBeEqualTo HttpStatusCode.BadRequest
+    }
+}
+
+fun TestApplicationEngine.testDeniedPersonAccess(
+    url: String,
+    validToken: String,
+    httpMethod: HttpMethod,
+) {
+    with(
+        handleRequest(httpMethod, url) {
+            addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
+            addHeader(
+                NAV_PERSONIDENT_HEADER,
+                UserConstants.PERSONIDENT_VEILEDER_NO_ACCESS.value
+            )
+        }
+    ) {
+        response.status() shouldBeEqualTo HttpStatusCode.Forbidden
+    }
+}
